// My Website/webpack---_N_E/src/components/Projects/Projects.types.ts
import type { ProjectDetails } from "../ProjectModal";

export enum ProjectsType {
  PROJECTS,
  LABS,
}

export type ProjectsProps = {
  maxCount?: number;
  projectsType?: ProjectsType;
};

export type ProjectProps = ProjectDetails & {
  title: string;
  titleTwo?: string;
  fullTitle: string;
  shortDescription: string;
  coverPhotoUrl: string;
};